Operation method attributes are moved down to the SenseNet.ContentRepository
