+++ src/ContentRepository/ApplicationModel/GrantedToGenericAction.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 
+// ReSharper disable once CheckNamespace
 namespace SenseNet.ApplicationModel
 {
     [AttributeUsage(AttributeTargets.Method)]
