+++ src/ContentRepository/ApplicationModel/ScenarioAttribute.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
+// ReSharper disable once CheckNamespace
 namespace SenseNet.ApplicationModel
 {
     /// <summary>
